fix: chrome precision bug in smoke shaders feat: add sine wave distortion to spark motion chore: add GPL license headers to particles

+++ src/engine/particles/point/shaders/fragment.frag
@@ -1,5 +1,28 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
 
-precision mediump float;
+/**
+*
+*    Particle fragment shader which takes texture as like smoke.
+*    Copyright (C) 2025  Tahsin Önemli
+*
+*    This program is free software: you can redistribute it and/or modify
+*    it under the terms of the GNU General Public License as published by
+*    the Free Software Foundation, either version 3 of the License, or
+*    (at your option) any later version.
+*
+*    This program is distributed in the hope that it will be useful,
+*    but WITHOUT ANY WARRANTY; without even the implied warranty of
+*    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+*    GNU General Public License for more details.
+*
+*    You should have received a copy of the GNU General Public License
+*    along with this program.  If not, see <https://www.gnu.org/licenses/>.
+*
+*/
+
+// Chrome-based browsers may produce visual glitches when using mediump precision in this shader.
+precision highp float;
+
 
 uniform sampler2D diffuseTexture;
